Ensure that CUBLAS and CuSOLVER streams are reset. Fix fMM bug.
Created a context manager for cuBLAS streams, similar to torch.cuda.stream
This was not sufficient to fix a synchronization bug in fMM (which only manifested
in certain cases: in the full InCoreFalkon run when the K_NM kernel went in memory).

To fix it, it was necessary to set a stream for _generic_fmm!