Added lauum_upper CUDA kernel
The upper LAUUM kernel was added (symmetric to the lower kernel) in lauum.cu
The corresponding code in the lower-c parallel_lauum.py function was adapted
and simplified to use this kernel. This gets rid of a GPU-GPU copy.

Additionally several parts of the parallel_lauum code were improved,
and a bug was fixed within the lower-f function when the `independent_output`
option was set.

Finally, tests were augmented for the new kernel, and to cover all
input possibilities for parallel_lauum.