Ensure storage_offset is int
Older versions of PyTorch complain if storage_offset is e.g. a np.int64
object (which it will be in certain cases, since it is the return value
of a np.prod call). We cast to int to make sure this is not a problem.

This should be done for size, stride and probably other arguments as well,
but since it's not a problem on newer PyTorch versions we keep things as
they are, since those are usually not np.int64.